MIG-36: refactoring variant group combination migration
